+++ labs/embeddings/vectorizers/vectorizer.py
